Former-commit-id: b61f18318d3da93d87d2c8b89115aa263b5582b2

+++ BE/arena/src/main/java/com/example/arena/domain/community/dto/response/DetailBoardResponse.java
@@ -14,5 +14,5 @@ import lombok.Setter;
 public class DetailBoardResponse {
 	private Board board;
 	private String memberNickName;
-	private int comments;
+	private long comments;
 }
